fix logging and add more trace logging
Abandoning keepalive feature for now because connections idling in
keepalive block new connections. We could not worry about it, reduce the
timeout (currently 60 seconds), or get fancy and close an idle connection in
keepalive when it is blocking an incoming connection. But the benefit of
keepalive is not so clear now.